Development (#7)
* Add timestamp to log messages

Log output now includes localtime in [YYYY-MM-DD HH:MM:SS] format.


* Add 4-byte ASN support (RFC 6793)

Negotiate 4-octet AS number capability in OPEN messages and parse
AS_PATH/AGGREGATOR attributes with 4-byte ASNs when both peers
support it. Uses AS_TRANS (23456) in the 2-byte OPEN field for
local ASNs > 65535.


* Add COMMUNITY path attribute parsing (RFC 1997)

Parse type code 8 COMMUNITY attributes as arrays of 4-byte values,
rendered as ASN:VALUE with well-known names (NO_EXPORT, NO_ADVERTISE,
NO_EXPORT_SUBCONFED).


* Add LARGE_COMMUNITY path attribute parsing (RFC 8092)


* Add include guards to bgp_print.h

Fixes redeclaration errors when the header is included via multiple
paths (bgp.h and bgp_peer.h both include bgp_print.h).


* Bump version to 0.0.8


---------

+++ src/bgp.h
@@ -13,10 +13,10 @@ struct bgp_peer;
 struct bgp_instance;
 struct bgp_local;
 
-struct bgp_instance *create_bgp_instance(uint16_t,  uint32_t, uint8_t);
+struct bgp_instance *create_bgp_instance(uint32_t,  uint32_t, uint8_t);
 void free_bgp_instance(struct bgp_instance *);
 
-unsigned int create_bgp_peer(struct bgp_instance *, const char *, const uint16_t, const char *);
+unsigned int create_bgp_peer(struct bgp_instance *, const char *, const uint32_t, const char *);
 unsigned int bgp_peer_source(struct bgp_instance *, unsigned int, const char *);
 
 int set_bgp_output(struct bgp_instance *, unsigned int, enum bgp_output);
